requirements.txt
enum34==1.1.6
namedlist==1.7
packaging==19.1
pbr==5.2.0
pluggy==0.6.0
py==1.8.0
pyparsing==2.4.2
pytest==3.6.3
pytest-allure-adaptor==1.7.10
pytest-forked==1.0.2
pytest-rerunfailures==5.0
pytest-xdist==1.23.2
PyYAML==3.13
selenium==3.141.0
